Bump wasi-sdk version to 19.0 (#1894)
1. Bump wasi-sdk version to 19.0

2. Use another way to work with sample/workloads
```
$ cd samples/workloads
$ cmake -S . -B build
$ cmake --build build
$ ctest -V --test-dir build
```
For now, it covers bwa, meshoptimizer and wasm-av1, but still
needs to run XNNPACK and tensorflow manually

3. Remove build-wasi-sdk
Since the latest wasi-sdk uses the latest llvm release, there is no
need to create a local wasi-sdk building. Only some headers of
emscripten-core are still required.
