+++ samples/workload/README.md
@@ -1,24 +1,13 @@
-All workloads have similar requirment of software dependencies, including
-**emsdk**, **wabt** and **binaryen**
+All workloads have similar requirment of software dependencies, including **emsdk** and **binaryen**
 
-> There might be slight differences when using MacOS and other Linux distro than Ubuntu. This document only target
-Ubuntu 20.04 as example.
+> There might be slight differences when using MacOS and other Linux distro than Ubuntu. This document targets
+Ubuntu 20.04 as an example.
 
 ## Installation instructions
 
-use [preparation.sh](./preparation.sh) to install all dependencies before compiling any workload.
+use [preparation.sh](./preparation.sh) to install all dependencies before compiling any workload. Or use [*vscode DevContainer*](../../.devcontainer/)
 
-for details, the script includes below steps:
-
-- **wabt**. Install
-  [latest release](https://github.com/WebAssembly/wabt/releases/download/1.0.31/wabt-1.0.31-ubuntu.tar.gz)
-  to */opt/wabt*
-
-``` bash
-$ wget https://github.com/WebAssembly/wabt/releases/download/${WABT_VER}/${WABT_FILE}
-$ tar zxf ${WABT_FILE} -C /opt
-$ ln -sf /opt/wabt-${WABT_VER} /opt/wabt
-```
+The script installs below software:
 
 - **emsdk**. Refer to [the guide](https://emscripten.org/docs/getting_started/downloads.html). Don't forget to activate
   emsdk and set up environment variables. Verify it with `echo ${EMSDK}`. Please be sure to install and activate the building
