+++ samples/workload/bwa/CMakeLists.bwa_wasm.txt
@@ -1,11 +1,14 @@
 # Copyright (C) 2019 Intel Corporation. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-cmake_minimum_required (VERSION 3.0)
+cmake_minimum_required (VERSION 3.14)
 
 project(bwa_wasm C)
 
-include(${CMAKE_CURRENT_SOURCE_DIR}/../../cmake/preparation.cmake)
+list(APPEND CMAKE_MODULE_PATH ${CMAKE_CURRENT_SOURCE_DIR}/../../cmake)
+
+################ dependencies ################
+find_package(Binaryen 111 REQUIRED)
 
 ################ LIBZ ################
 set(LIBZ_SRC_DIR ${CMAKE_CURRENT_SOURCE_DIR}/../libz)
@@ -86,12 +89,6 @@ add_executable(${PROJECT_NAME} ${BWA_SOURCE})
 
 set_target_properties(${PROJECT_NAME} PROPERTIES OUTPUT_NAME bwa.wasm)
 
-target_include_directories(${PROJECT_NAME}
-  PRIVATE
-    ${WASI_SDK_HOME}/share/wasi-sysroot/include/libc/musl
-    ${WASI_SDK_HOME}/share/wasi-sysroot/include/sse
-)
-
 target_compile_definitions(${PROJECT_NAME}
   PRIVATE
     USE_MALLOC_WRAPPERS
@@ -117,7 +114,7 @@ target_link_libraries(${PROJECT_NAME} z_wasm wasi-emulated-process-clocks)
 
 add_custom_target(bwa_wasm_opt ALL
   COMMAND
-    ${WASM_OPT} -Oz --enable-simd -o bwa.opt.wasm bwa.wasm
+    ${Binaryen_WASM_OPT} -Oz --enable-simd -o bwa.opt.wasm bwa.wasm
   BYPRODUCTS
     ${CMAKE_CURRENT_BINARY_DIR}/bwa.opt.wasm
   WORKING_DIRECTORY
