+++ samples/workload/bwa/CMakeLists.txt
@@ -1,11 +1,14 @@
 # Copyright (C) 2019 Intel Corporation. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-cmake_minimum_required (VERSION 2.8...3.16)
+cmake_minimum_required (VERSION 3.14)
 
 project(bwa_wasm)
 
-include(${CMAKE_CURRENT_SOURCE_DIR}/../cmake/preparation.cmake)
+list(APPEND CMAKE_MODULE_PATH ${CMAKE_CURRENT_SOURCE_DIR}/../cmake)
+
+################ dependencies ################
+find_package(WASISDK 16.0 REQUIRED)
 
 #######################################
 include(ExternalProject)
@@ -13,7 +16,7 @@ include(ExternalProject)
 ################ libz ################
 ExternalProject_Add(libz_src
   GIT_REPOSITORY https://github.com/madler/zlib.git
-  GIT_TAG        master
+  GIT_TAG        04f42ceca40f73e2978b50e93806c2a18c1281fc
   GIT_PROGRESS   ON
   GIT_SHALLOW    ON
   SOURCE_DIR     ${CMAKE_CURRENT_SOURCE_DIR}/libz
@@ -27,7 +30,7 @@ ExternalProject_Add(libz_src
 ################ bwa ################
 ExternalProject_Add(bwa
   GIT_REPOSITORY https://github.com/lh3/bwa.git
-  GIT_TAG        master
+  GIT_TAG        139f68fc4c3747813783a488aef2adc86626b01b
   GIT_PROGRESS   ON
   GIT_SHALLOW    ON
   SOURCE_DIR     ${CMAKE_CURRENT_SOURCE_DIR}/bwa
@@ -37,10 +40,30 @@ ExternalProject_Add(bwa
                   && ${CMAKE_COMMAND} -E copy ${CMAKE_CURRENT_SOURCE_DIR}/CMakeLists.bwa_wasm.txt CMakeLists.txt
                   && git apply ../bwa.patch
   CONFIGURE_COMMAND  ${CMAKE_COMMAND}
-                       -DWASI_SDK_PREFIX=${WASI_SDK_HOME}
-                       -DCMAKE_TOOLCHAIN_FILE=${WASI_SDK_HOME}/share/cmake/wasi-sdk.cmake
-                       -DCMAKE_SYSROOT=${WASI_SDK_HOME}/share/wasi-sysroot
+                       -DWASI_SDK_PREFIX=${WASISDK_HOME}
+                       -DCMAKE_TOOLCHAIN_FILE=${WASISDK_TOOLCHAIN}
+                       -DCMAKE_SYSROOT=${WASISDK_SYSROOT}
+                       -DCMAKE_C_FLAGS=-isystem\ ${CMAKE_CURRENT_SOURCE_DIR}/../include/sse\ -isystem\ ${CMAKE_CURRENT_SOURCE_DIR}/../include/libc/musl
+                       --debug-output --debug-find
                        ${CMAKE_CURRENT_SOURCE_DIR}/bwa
   BUILD_COMMAND  make bwa_wasm_opt
-  INSTALL_COMMAND ${CMAKE_COMMAND} -E copy ./bwa.opt.wasm ${CMAKE_BINARY_DIR}/bwa.wasm
+  INSTALL_COMMAND ${CMAKE_COMMAND} -E copy_if_different ./bwa.opt.wasm ${CMAKE_CURRENT_BINARY_DIR}/bwa.wasm
+)
+
+################ bwa data ################
+ExternalProject_Add(bwa-kit
+  PREFIX            bwa-kit
+  URL               https://sourceforge.net/projects/bio-bwa/files/bwakit/bwakit-0.7.15_x64-linux.tar.bz2/download
+  URL_HASH          SHA256=0a7b11971bc7916b68e9df35a364afe77cb3000df02ffb3a6fbd1aff9be5878c
+  DOWNLOAD_EXTRACT_TIMESTAMP   ON
+  DOWNLOAD_NAME     bwakit-0.7.15_x64-linux.tar.bz2
+  DOWNLOAD_NO_EXTRACT          OFF
+  DOWNLOAD_NO_PROGRESS         ON
+  UPDATE_COMMAND    ""
+  PATCH_COMMAND     ""
+  CONFIGURE_COMMAND ""
+  BUILD_COMMAND     ""
+  INSTALL_COMMAND   ${CMAKE_COMMAND} -E copy_if_different
+                      ${CMAKE_CURRENT_BINARY_DIR}/bwa-kit/src/bwa-kit/resource-GRCh38/hs38DH-extra.fa
+                      ${CMAKE_CURRENT_BINARY_DIR}/hs38DH-extra.fa
 )
