+++ samples/workload/wasm-av1/CMakeLists.avx_wasm.txt
@@ -1,14 +1,14 @@
 # Copyright (C) 2019 Intel Corporation. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-cmake_minimum_required (VERSION 2.8...3.16)
+cmake_minimum_required (VERSION 3.14)
 
 project(testavx)
 
-include(${CMAKE_CURRENT_SOURCE_DIR}/../../cmake/preparation.cmake)
+list(APPEND CMAKE_MODULE_PATH ${CMAKE_CURRENT_SOURCE_DIR}/../../cmake)
 
-# a workaround to let aom find our non-public headers
-include_directories(${WASI_SDK_HOME}/share/wasi-sysroot/include/libc/musl)
+################ dependencies ################
+find_package(Binaryen 111 REQUIRED)
 
 ################  AOM  ################
 set(ENABLE_CCACHE ON)
@@ -62,7 +62,7 @@ add_dependencies(${PROJECT_NAME} aom)
 
 add_custom_target(${PROJECT_NAME}_opt ALL
   COMMAND
-    ${WASM_OPT} -Oz --enable-simd -o ${PROJECT_NAME}.opt.wasm ${PROJECT_NAME}.wasm
+    ${Binaryen_WASM_OPT} -Oz --enable-simd -o ${PROJECT_NAME}.opt.wasm ${PROJECT_NAME}.wasm
   BYPRODUCTS
     ${CMAKE_CURRENT_BINARY_DIR}/${PROJECT_NAME}.opt.wasm
   WORKING_DIRECTORY
