+++ samples/workload/wasm-av1/CMakeLists.txt
@@ -1,11 +1,14 @@
 # Copyright (C) 2019 Intel Corporation. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-cmake_minimum_required (VERSION 2.8...3.16)
+cmake_minimum_required (VERSION 3.14)
 
 project(av1_wasm)
 
-include(${CMAKE_CURRENT_SOURCE_DIR}/../cmake/preparation.cmake)
+list(APPEND CMAKE_MODULE_PATH ${CMAKE_CURRENT_SOURCE_DIR}/../cmake)
+
+################ dependencies ################
+find_package(WASISDK 16.0 REQUIRED)
 
 #######################################
 include(ExternalProject)
@@ -23,10 +26,14 @@ ExternalProject_Add(av1
                   && ${CMAKE_COMMAND} -E copy ${CMAKE_CURRENT_SOURCE_DIR}/CMakeLists.avx_wasm.txt CMakeLists.txt
                   && git apply ../av1-clang.patch
   CONFIGURE_COMMAND  ${CMAKE_COMMAND}
-                      -DWASI_SDK_PREFIX=${WASI_SDK_HOME}
-                      -DCMAKE_TOOLCHAIN_FILE=${WASI_SDK_HOME}/share/cmake/wasi-sdk.cmake
-                      -DCMAKE_SYSROOT=${WASI_SDK_HOME}/share/wasi-sysroot
+                      -DWASI_SDK_PREFIX=${WASISDK_HOME}
+                      -DCMAKE_TOOLCHAIN_FILE=${WASISDK_TOOLCHAIN}
+                      -DCMAKE_SYSROOT=${WASISDK_SYSROOT}
+                       -DCMAKE_C_FLAGS=-isystem\ ${CMAKE_CURRENT_SOURCE_DIR}/../include/sse\ -isystem\ ${CMAKE_CURRENT_SOURCE_DIR}/../include/libc/musl
                       ${CMAKE_CURRENT_SOURCE_DIR}/av1
   BUILD_COMMAND     make testavx_opt
-  INSTALL_COMMAND   ${CMAKE_COMMAND} -E copy testavx.opt.wasm ${CMAKE_CURRENT_BINARY_DIR}/testavx.wasm
+  INSTALL_COMMAND   ${CMAKE_COMMAND} -E copy_if_different
+                      testavx.opt.wasm
+                      ${CMAKE_CURRENT_SOURCE_DIR}/av1/third_party/samples/elephants_dream_480p24.ivf
+                      ${CMAKE_CURRENT_BINARY_DIR}
 )
