--- test-tools/build-wasi-sdk/patches/wasi_sdk.patch
@@ -1,17 +1,0 @@
-diff --git a/version.sh b/version.sh
-index 8e7c44c..ff0d3ba 100755
---- a/version.sh
-+++ b/version.sh
-@@ -1,6 +1,7 @@
- #!/usr/bin/env bash
--set -e
--git config --global --add safe.directory "/workspace"
--GIT_DESCR=$(git describe --long --candidates=999 --match='wasi-sdk-*' --dirty='+m' --abbrev=12)
--GIT_PACKAGE_VERSION=$(echo $GIT_DESCR | perl -ne 'if(/^wasi-sdk-(\d+)-(\d+)-g([0-9a-f]{7,12})([+]m)?$/) { if($2 == 0) { print "$1.$2$4" } else { print "$1.$2g$3$4" } exit } else { print "could not parse git description"; exit 1 }';)
--echo $GIT_PACKAGE_VERSION
-+#set -e
-+#git config --global --add safe.directory "/workspace"
-+#GIT_DESCR=$(git describe --long --candidates=999 --match='wasi-sdk-*' --dirty='+m' --abbrev=12)
-+#GIT_PACKAGE_VERSION=$(echo $GIT_DESCR | perl -ne 'if(/^wasi-sdk-(\d+)-(\d+)-g([0-9a-f]{7,12})([+]m)?$/) { if($2 == 0) { print "$1.$2$4" } else { print "$1.$2g$3$4" } exit } else { print "could not parse git description"; exit 1 }';)
-+#echo $GIT_PACKAGE_VERSION
-+echo wasi-sdk-17-eng
